requirements.txt
absl-py==0.13.0
addict==2.4.0
antlr4-python3-runtime==4.8
argon2-cffi==20.1.0
async-generator==1.10
attrs==21.2.0
backcall==0.2.0
bleach==3.3.1
cached-property==1.5.2
cachetools==4.2.2
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.3
click==8.0.1
configparser==5.0.2
cycler==0.10.0
debugpy==1.4.1
decorator==4.4.2
defusedxml==0.7.1
docker-pycreds==0.4.0
entrypoints==0.3
filelock==3.0.12
gdown==3.13.0
gitdb==4.0.7
GitPython==3.1.18
google-auth==1.33.1
google-auth-oauthlib==0.4.4
googledrivedownloader==0.4
gql==0.2.0
graphql-core==1.1
grpcio==1.39.0
h5py==3.3.0
hydra-core==1.1.0
idna==3.2
imageio==2.9.0
importlib-metadata==3.10.1
importlib-resources==5.2.0
ipykernel==6.0.3
ipython==7.25.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
isodate==0.6.0
jedi==0.18.0
Jinja2==3.0.1
joblib==1.0.1
jsonpatch==1.32
jsonpointer==2.1
jsonschema==3.2.0
jupyter-client==6.1.12
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
kiwisolver==1.3.1
llvmlite==0.33.0
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib==3.4.2
matplotlib-inline==0.1.2
MinkowskiEngine==0.5.4
mistune==0.8.4
nbclient==0.5.3
nbconvert==6.1.0
nbformat==5.1.3
nest-asyncio==1.5.1
networkx==2.5.1
ninja==1.10.0.post3
notebook==6.4.0
numba==0.50.1
numpy==1.19.5
nvidia-ml-py3==7.352.0
oauthlib==3.1.1
omegaconf==2.1.0
open3d==0.12.0
openmesh==1.1.6
packaging==21.0
pandas==1.3.1
pandocfilters==1.4.3
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.3.1
plyfile==0.7.4
prometheus-client==0.11.0
promise==2.3
prompt-toolkit==3.0.19
protobuf==3.17.3
psutil==5.8.0
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
Pygments==2.9.0
pyparsing==2.4.7
pyrsistent==0.18.0
PySocks==1.7.1
python-dateutil==2.8.2
python-louvain==0.15
pytorch-metric-learning==0.9.99
pytz==2021.1
PyWavelets==1.1.1
PyYAML==5.4.1
pyzmq==22.1.0
rdflib==6.0.0
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
scikit-image==0.16.2
scikit-learn==0.24.2
scipy==1.7.0
Send2Trash==1.7.1
sentry-sdk==1.3.0
shortuuid==1.0.1
six==1.16.0
sklearn==0.0
smmap==4.0.0
subprocess32==3.5.4
tensorboard==2.5.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
terminado==0.10.1
testpath==0.5.0
threadpoolctl==2.2.0
torch==1.9.0+cu111
torch-cluster==1.5.9
torch-geometric==1.7.2
torch-points-kernels==0.7.0
torch-points3d @ git+https://github.com/nicolas-chaulet/torch-points3d.git@5378a2ae7ef612433acedaa497073971ec66ff09
torch-scatter==2.0.7
torch-sparse==0.6.10
torch-spline-conv==1.2.1
torchfile==0.1.0
torchnet==0.0.4
torchvision==0.10.0+cu111
tornado==6.1
tqdm==4.61.2
traitlets==5.0.5
types-requests==0.1.13
types-six==0.1.7
typing-extensions==3.10.0.0
urllib3==1.26.6
visdom==0.1.8.9
wandb==0.8.36
watchdog==2.1.3
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.1.0
Werkzeug==2.0.1
widgetsnbextension==3.5.1
zipp==3.5.0
